gradio_demo_app/requirements.txt
torch==2.0.1+cpu
opencv-python==4.8.0.76
torchvision==0.15.2+cpu
tqdm==4.66.2
Pillow==10.0.0
matplotlib==3.7.2
ultralytics==8.1.11
numpy==1.25.2
pandas==2.0.3
gradio==4.25.0
supervision==0.17.1
transformers==4.33.1